+++ app/admin/image_series.rb
@@ -501,64 +501,65 @@ ActiveAdmin.register ImageSeries do
     link_to('DICOM Metadata', dicom_metadata_admin_image_series_path(resource)) if resource.has_dicom? && can?(:read_dicom_metadata, resource)
   end
 
-  collection_action :batch_assign_to_patient, :method => :post do
-    if(params[:patient_id].nil?)
-      flash[:error] = 'You have to select a patient to assign these image series to.'
-      redirect_back(fallback_location: admin_image_series_index_path)
-      return
-    elsif(params[:image_series].nil?)
-      flash[:error] = 'You have to specify at least one image series to assign to this patient.'
-      redirect_back(fallback_location: admin_image_series_index_path)
-      return
-    end
+  # TODO: Remove batch assign to patient. See: #5783
+  # collection_action :batch_assign_to_patient, :method => :post do
+  #   if(params[:patient_id].nil?)
+  #     flash[:error] = 'You have to select a patient to assign these image series to.'
+  #     redirect_back(fallback_location: admin_image_series_index_path)
+  #     return
+  #   elsif(params[:image_series].nil?)
+  #     flash[:error] = 'You have to specify at least one image series to assign to this patient.'
+  #     redirect_back(fallback_location: admin_image_series_index_path)
+  #     return
+  #   end
 
-    image_series_ids = params[:image_series].split(' ')
-    image_series = ImageSeries.find(image_series_ids)
+  #   image_series_ids = params[:image_series].split(' ')
+  #   image_series = ImageSeries.find(image_series_ids)
 
-    patient = Patient.find(params[:patient_id])
-    authorize! :read, patient
+  #   patient = Patient.find(params[:patient_id])
+  #   authorize! :read, patient
 
-    image_series.each do |i_s|
-      authorize! :assign_patient, i_s
-      next unless i_s.visit_id.nil?
+  #   image_series.each do |i_s|
+  #     authorize! :assign_patient, i_s
+  #     next unless i_s.visit_id.nil?
 
-      i_s.patient = patient
-      unless(i_s.series_number.blank? or i_s.patient.image_series.where(:series_number => i_s.series_number).empty?)
-        i_s.series_number = nil
-      end
-      i_s.save
-    end
+  #     i_s.patient = patient
+  #     unless(i_s.series_number.blank? or i_s.patient.image_series.where(:series_number => i_s.series_number).empty?)
+  #       i_s.series_number = nil
+  #     end
+  #     i_s.save
+  #   end
 
-    redirect_to params[:return_url], :notice => 'The image series were assigned to the patient.'
-  end
+  #   redirect_to params[:return_url], :notice => 'The image series were assigned to the patient.'
+  # end
 
-  batch_action :assign_to_patient, confirm: 'This will modify all selected image series. Are you sure?', if: proc { can? :assign_patient, ImageSeries } do |selection|
-    failure = false
-    study_id = nil
+  # batch_action :assign_to_patient, confirm: 'This will modify all selected image series. Are you sure?', if: proc { can? :assign_patient, ImageSeries } do |selection|
+  #   failure = false
+  #   study_id = nil
 
-    ImageSeries.find(selection).each do |image_series|
-      study_id = image_series.study.id if study_id.nil?
+  #   ImageSeries.find(selection).each do |image_series|
+  #     study_id = image_series.study.id if study_id.nil?
 
-      if(image_series.study.id != study_id)
-        flash[:error] = 'Not all selected image series belong to the same study. Batch assignment can only be used for series of the same study.'
-        redirect_back(fallback_location: admin_image_series_index_path)
-        failure = true
-        break
-      elsif(image_series.visit_id != nil)
-        flash[:error] = 'Not all selected image series are currently unassigned. Batch assignment can only be used for series which are not currently assigned to a visit.'
-        redirect_back(fallback_location: admin_image_series_index_path)
-        failure = true
-        break
-      end
-    end
-    next if failure
+  #     if(image_series.study.id != study_id)
+  #       flash[:error] = 'Not all selected image series belong to the same study. Batch assignment can only be used for series of the same study.'
+  #       redirect_back(fallback_location: admin_image_series_index_path)
+  #       failure = true
+  #       break
+  #     elsif(image_series.visit_id != nil)
+  #       flash[:error] = 'Not all selected image series are currently unassigned. Batch assignment can only be used for series which are not currently assigned to a visit.'
+  #       redirect_back(fallback_location: admin_image_series_index_path)
+  #       failure = true
+  #       break
+  #     end
+  #   end
+  #   next if failure
 
-    @return_url = request.referer
-    patients = Study.find(study_id).patients
+  #   @return_url = request.referer
+  #   patients = Study.find(study_id).patients
 
-    @page_title = 'Assign to Patient'
-    render 'admin/image_series/assign_to_patient', :locals => {:selection => selection, :patients => patients}
-  end
+  #   @page_title = 'Assign to Patient'
+  #   render 'admin/image_series/assign_to_patient', :locals => {:selection => selection, :patients => patients}
+  # end
 
   collection_action :batch_assign_to_visit, :method => :post do
     if(params[:visit_id].blank?)