+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Patches / Fixes
 
 * Fix - #5780 - Fix inconsistent image series form behavior when changing patient or visit, by removing the patient and visit fields
+* Fix - #5783 - Remove "Batch assign image series to patient" as we currently do not adjust DICOM metadata automatically
 
 ## 7.2.1